# api/nodejs-server/api/swagger.yaml
swagger: "2.0"
info:
  description: "Provides a RESTful API for interacting with Smartsplit. Documentation is written in both French and English."
  version: "v1"
  title: "Smartsplit API"
  x-origin:
  - format: "yaml"
    url: "https://raw.githubusercontent.com/iptoki/smartsplit/develop/api/swagger.yaml"
  x-providerName: "smartsplit.org"
  x-logo:
    url: "http://smartsplit.org/wp-content/uploads/2019/05/Logo-v02-005-noir.png"
#host: api.smartsplit.org
basePath: "/v1"
...
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- jwt: []
paths:
...
  /proposal:
    get:
      tags:
      - "Proposals"
      description: "Get a list of all media split proposals"
      operationId: "getAllProposals"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/listProposals"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "Proposals"
    post:
      tags:
      - "Proposals"
      description: "This method creates a new split proposal for a given media"
      operationId: "postProposal"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/proposal"
      responses:
        200:
          description: "OK. The split was created."
          schema:
            $ref: "#/definitions/proposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
...
  /proposal/derniere-proposition/{mediaId}:
    get:
      tags:
      - "Proposals"
      description: "La dernière proposition pour un média"
      operationId: "getDernierePropositionPourMedia"
      parameters:
      - name: "mediaId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/proposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "Proposals"
  /proposal/media/{mediaId}:
    get:
      tags:
      - "Proposals"
      description: "Get a list of all split proposals for a given mediaId"
      operationId: "getMediaProposals"
      parameters:
      - name: "mediaId"
        in: "path"
        description: "The media unique ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/listProposals"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "Proposals"
  /proposal/invite:
    post:
      tags:
      - Proposals
      description: "This method invite a right holder to a proposal"
      operationId: "inviteProposal"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/inviteProposal"
      responses:
        200:
          description: "OK."
          schema:
            $ref: "#/definitions/inviteProposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/justifier-refus:
    post:
      tags:
      - Proposals
      description: "This method justify a split decline"
      operationId: "justifierRefus"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/justifierRefus"
      responses:
        200:
          description: "OK."
          schema:
            $ref: "#/definitions/justifierRefus"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/liste-votes:    
    post:
      tags:
      - Proposals
      description: "This method returns the votation state for a proposal"
      operationId: "listeVotes"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/listeVote"
      responses:
        200:
          description: "OK."
          schema:
            $ref: "#/definitions/listeVote"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/decode:    
    post:
      tags:
      - Proposals
      description: "This method decode a vote token for a proposal"
      operationId: "decodeProposal"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/decodeProposal"
      responses:
        200:
          description: "OK."
          schema:
            $ref: "#/definitions/decodeProposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"  
  /proposal/voter:    
    post:
      tags:
      - "Splits"
      description: "This method accepts a vote for a proposal"
      operationId: "voteProposal"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/voteProposal"
      responses:
        200:
          description: "OK."
          schema:
            $ref: "#/definitions/voteProposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/{uuid}:
    get:
      tags:
      - "Proposals"
      description: "Get a split proposal with the given ID"
      operationId: "getProposal"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split proposal's unique ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/proposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "Proposals"
    put:
      tags:
      - "Proposals"
      description: "This method updates a split proposal"
      operationId: "updateProposal"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split proposal's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/proposal"
      responses:
        200:
          description: "OK. The split proposal was created."
          schema:
            $ref: "#/definitions/proposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
    delete:
      tags:
      - "Proposals"
      description: "Delete a right split proposal with the given ID"
      operationId: "deleteProposal"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The splits proposal's unique profile ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK. Split proposal deleted."
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/rightHolder/{rightHolderId}:
    get:
      tags:
      - "Proposals"
      description: "Get all split proposals for the given right holder"
      operationId: "getProposalsRightHolder"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The Id of the given right holder"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/proposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "Proposals"
  /proposal/{uuid}/mediaId:
    patch:
      tags:
      - "Proposals"
      description: "Update mediaId for given split proposal"
      operationId: "patchProposalMediaId"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split proposal's unique ID"
        required: true
        type: "string"
      - in: "body"
        name: "mediaId"
        description: "The split proposal's media Id"
        required: true
        schema:
          $ref: "#/definitions/proposal/properties/mediaId"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              mediaId:
                type: "integer"
                example: 1
                description: "The mediaId for the given split"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/{uuid}/initiator:
    patch:
      tags:
      - "Proposals"
      description: "Update initiator for a given split proposal"
      operationId: "patchProposalInitiator"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split's unique ID"
        required: true
        type: "string"
      - in: "body"
        name: "initiator"
        description: "The initiator of the given split proposal"
        required: true
        schema:
          $ref: "#/definitions/proposal/properties/initiator"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              splitType:
                type: "object"
                example: {"rightHolderId": 1, "name": "Joe Smith"}
                description: "The initator object of the given split proposal"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/{uuid}/rightsSplits:
    patch:
      tags:
      - "Proposals"
      description: "Update rights split object for given split proposal"
      operationId: "patchProposalRightsSplits"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split proposal's unique ID"
        required: true
        type: "string"
      - in: "body"
        name: "rightsSplits"
        description: "The split proposal's rights splits object"
        required: true
        schema:
          $ref: "#/definitions/proposal/properties/rightsSplits"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/proposal/properties/rightsSplits"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
    post:
      tags:
      - "Proposals"
      description: "Create rights split smart contracts for given split proposal"
      operationId: "postProposalRightsSplits"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split proposal's unique ID"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "The split proposal's rights splits object"
        required: true
        schema:
          type: "object"
          properties: 
            uuid:
              type: "string"
              example: "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e"
              description: "The unique ID for the given split proposal"
            mediaId:
              type: "integer"
              example: 1
              description: "The Id for the given media which corresponds to the right split"
            rightsSplits:
              type: "object"
              additionalProperties:
                type: "object"
                description: "the copyright object for the given right split"
            initiatorUuid: 
              type: "string"
            initiatorName: 
              type: "string"              
            creationDate:
              type: "string"
              format: "iso8601-date-time"
              example: "2019-01-01T15:53:00"
              description: "the creation date generated at time of split proposal creation"
            _d:
              type: "string"
              example: "1566229239"
              description: "the creation date of split stored as a Unix timestamp (in seconds)"
            comments: 
              type: "array"
              items:
                type: "object"
                description: "the comment and rightHolderId for the given proposal"
            state:
              type: "string"
              description: "define the state of the proposal (BROUILLON, PRET, VOTATION, ACCEPTE, REFUSE"
          example:
            {
              "uuid": "12345c60-7b1a-11e8-9c9c-2d42b21b1a3e",
              "mediaId": "1",
              "initiatorUuid": "40e5ad76-d2c6-4f1e-9eb4-98213e29e36f", 
              "initiatorName": "Jim Smith",
              "creationDate": "2019-01-01T15:53:00",
              "_d": "1566229239",
              "rightsSplits": {
                "workCopyrightSplit": {                                       
                  "music": [
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a3e": "songwriter",
                      "45745c60-7b1a-11e8-9c9c-2d42b21b1a3i": "composer"
                    },
                    "rightHolder": {
                      "rightHolderId": "40e5ad76-d2c6-4f1e-9eb4-98213e29e36f",
                      "name": "Joe Smith",
                      "color": "#FFFFFFF"
                    },
                    "splitPct": "50",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:45:51Z"
                  },
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a3f": "songwriter",
                      "45745c60-7b1a-11e8-9c9c-2d42b21b1a3i": "composer"
                    },
                    "rightHolder": {
                      "rightHolderId": "523bac7a-b14c-4a33-87ad-6aa3e848f2f6",
                      "name": "Bob Andrews",
                      "color": "#FFFFFFF"
                    },
                    "splitPct": "25",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:46:51Z"
                  }
                  ],
                  "lyrics": [
                    {
                      "contributorRole": {
                        "12345c60-7b1a-11e8-9c9c-2d42b21b1a4g": "arranger"
                      },
                      "rightHolder": {
                        "rightHolderId": "8c9a463d-4803-4221-9e96-f940f250e87d",
                        "name": "Joe Duchane",
                        "color": "#000000"
                      },    
                      "splitPct": "25",
                      "voteStatus": "active",
                      "_t": "2019-07-08T16:47:51Z"
                    }
                  ]
                },
                "performanceNeighboringRightSplit": {
                "principal": [
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a3e": "guitarist",
                      "45745c60-7b1a-11e8-9c9c-2d42b21b1a3i": "writer"
                    },
                    "rightHolder": {
                      "rightHolerId": "40e5ad76-d2c6-4f1e-9eb4-98213e29e36f",
                      "name": "Joe Smith",
                      "color": "#111111"
                    },
                    "splitPct": "80",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:40:51Z"
                  }
                ],
                "accompaniment": [
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a3f": "flutist",
                      "45745c60-7b1a-11e8-9c9c-2d42b21b1a3i": "writer"
                    },
                    "rightHolder": {
                      "rightHolderId": "523bac7a-b14c-4a33-87ad-6aa3e848f2f6",
                      "name": "Bob Andrews",
                      "color": "#222222"
                    },
                    "splitPct": "20",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:39:51Z"
                  }
                ]
                },
                "masterNeighboringRightSplit": [
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a3e": "guitarist",
                      "45745c60-7b1a-11e8-9c9c-2d42b21b1a3i": "writer"
                    },
                    "rightHolder": {
                      "rightHolderId": "40e5ad76-d2c6-4f1e-9eb4-98213e29e36f",
                      "name": "Joe Smith",
                      "color": "#222222"
                    },
                    "splitPct": "50",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:45:51Z"
                  },
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a3f": "flutist",
                      "45745c60-7b1a-11e8-9c9c-2d42b21b1a3i": "writer"
                    },
                    "rightHolder": {
                      "rightHolderId": "523bac7a-b14c-4a33-87ad-6aa3e848f2f6",
                      "name": "Bob Andrews",
                      "color": "#555555"
                    },
                    "splitPct": "25",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:46:51Z"
                  },
                  {
                    "contributorRole": {
                      "12345c60-7b1a-11e8-9c9c-2d42b21b1a4g": "composer"
                    },
                    "rightHolder": {
                      "rightHolderId": "8c9a463d-4803-4221-9e96-f940f250e87d",
                      "name": "Joe Duchane",
                      "color": "#FFFFFF"
                    },
                    "splitPct": "25",
                    "voteStatus": "active",
                    "_t": "2019-07-08T16:47:51Z"
                  }
                ]
              },
              "comments": [{
                "rightHolderId": "40e5ad76-d2c6-4f1e-9eb4-98213e29e36f",
                "comment": "Ok"
              },
              {
                "rightHolderId": "523bac7a-b14c-4a33-87ad-6aa3e848f2f6",
                "comment": "Not enough, more split"
              }
              ]
          }
      responses:
        200:
          description: "Smart contracts created"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
  /proposal/{uuid}/comments:
    patch:
      tags:
      - "Proposals"
      description: "Update coomments array for given split proposal"
      operationId: "patchProposalComments"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The split proposal's unique ID"
        required: true
        type: "string"
      - in: "body"
        name: "comments"
        description: "The split proposal's comments array"
        required: true
        schema:
          $ref: "#/definitions/proposal/properties/comments"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/proposal/properties/comments"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "Proposals"
...
securityDefinitions:
  basicAuth:
    description: "HTTP Basic Authentication"
    type: "basic"
  jwt:
    description: "A JSON Web Token (JWT) used for access control and authorization"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  listProposals:
    type: "array"
    description: "An array of information about splits"
    items:
      $ref: "#/definitions/proposal"
  justifierRefus:
    type: "object"
    properties:
      userId:
        type: "string"
        example: "1"
        description: "UUID utilisateur"      
      jeton:
        type: "string"
        example: "JWT Token"
        description: "Le jeton JWT à décoder"
      raison:
        type: "string"
        example: "Je ne suis pas d'accord"
        description: "Raison du refus"
  rightHolderShort:
    type: "object"
    description: "Structure courte d'un ayant-droit (nom, id, courriel)"
    properties:
      name:
        type: "string"
      rightHolderId:
        type: "string"
      email:
        type: "string"
  inviteProposal:
    type: "object"
    properties: 
      proposalId:
        type: "string"
        example: "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e"
        description: "The unique ID for the given split"
      rightHolders:
        type: "object"
        example: "{{rightHolderId: '2', name: 'Vincent', email: 'vincent@smartsplit.org'}, ... }"
        description: "Un tableau de colaborateurs (nom, id et courriel)"
        items:
          $ref: "#/definitions/rightHolderShort"
    example:
      {
        "proposalId": "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e"
      }
  listeVote:
    type: "object"
    properties: 
      proposalId:
        type: "string"
        example: "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e"
        description: "UUID de la propositino"
    example:
      {
        "splitId": "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e.xyz.xzdv"
      }
  decodeProposal:
    type: "object"
    properties: 
      jeton:
        type: "string"
        example: "JWT Token"
        description: "Le jeton JWT à décoder"      
    example:
      {
        "jeton": "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e.xyz.xzdv"
      }
  voteProposal:
    type: "object"
    properties:
      userId:
        type: "string"
        example: "1"
        description: "UUID utilisateur"
      jeton:
        type: "string"
        example: "JWT Token"
        description: "Le jeton JWT à décoder"
      droits:
        type: "object"
        additionalProperties:
          type: "object"
        example: "{workCopyrightSplit: 'ACCEPTE', ...}"
        description: "Objet décrivant les splits et les états"
    example:
      {
        "jeton": "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e.xyz.xzdv",
        "userId": "1",
        "droit": "workCopyrightSplit"
      }
  proposal:
    type: "object"
    properties: 
      uuid:
        type: "string"
        example: "45745c60-7b1a-11e8-9c9c-2d42b21b1a3e"
        description: "The unique ID for the given split proposal"
      mediaId:
        type: "integer"
        example: 1
        description: "The Id for the given media which corresponds to the right split"
      rightsSplits:
        type: "object"
        additionalProperties:
          type: "object"
          description: "the copyright object for the given right split"
      initiatorUuid: 
        type: "string"        
      initiatorName: 
        type: "string"
      creationDate:
        type: "string"
        format: "iso8601-date-time"
        example: "2019-01-01T15:53:00"
        description: "the creation date generated at time of split proposal creation"
      _d:
        type: "string"
        example: "1566229239"
        description: "the creation date of split stored as a Unix timestamp (in seconds)"
      comments: 
        type: "array"
        items:
          type: "object"
          description: "the comment and rightHolderId for the given proposal"
      state:
        type: "string"
        description: "define the state of the proposal (BROUILLON, PRET, VOTATION, ACCEPTE, REFUSE"    
...